+++ packages/pigeon/platform_tests/test_plugin/darwin/test_plugin/Package.swift
@@ -17,7 +17,7 @@ let package = Package(
     .library(name: "test-plugin", targets: ["test_plugin"])
   ],
   dependencies: [],
-  // #docregion spm-targets
+  // #docregion swiftpm-targets
   targets: [
     .target(
       name: "test_plugin_objc_gen",
@@ -29,5 +29,5 @@ let package = Package(
       dependencies: ["test_plugin_objc_gen"]
     ),
   ]
-  // #enddocregion spm-targets
+  // #enddocregion swiftpm-targets
 )