+++ packages/pigeon/platform_tests/test_plugin/example/jnigen_config.dart
@@ -14,7 +14,6 @@ void main() async {
       summarizerOptions: SummarizerOptions(backend: SummarizerBackend.asm),
       outputConfig: OutputConfig(
         dartConfig: DartCodeOutputConfig(
-          // Path is relative to appDirectory.
           path: Uri.file('../../shared_test_plugin_code/lib/src/generated/ni_tests.gen.jni.dart'),
           structure: OutputStructure.singleFile,
         ),